perf(compress、lang、poi): 代码优化

+++ pangju-commons-lang/src/main/java/io/github/pangju666/commons/lang/utils/UUIDUtils.java
@@ -29,7 +29,7 @@ public class UUIDUtils {
 		randomBytes[6] &= 0x0f;  /* clear version        */
 		randomBytes[6] |= 0x40;  /* set to version 4     */
 		randomBytes[8] &= 0x3f;  /* clear variant        */
-		randomBytes[8] |= 0x80;  /* set to IETF variant  */
+		randomBytes[8] |= (byte) 0x80;  /* set to IETF variant  */
 
 		long msb = 0;
 		long lsb = 0;
